+++ R/color.R
@@ -1,15 +1,16 @@
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
-#' @param x PARAM_DESCRIPTION
-#' @param alpha PARAM_DESCRIPTION, Default: FALSE
-#' @return OUTPUT_DESCRIPTION
-#' @details DETAILS
+#' @title
+#' Color to Hex
+#' @description
+#' Get the hex code for a color given by its name.
+#'
+#' @inheritParams grDevices::col2rgb
+#' @param x Color name as string.
+#' @return
+#' Vector of length 1 or greater of hex codes matched to the RGB derived
+#' from the color name.
+#'
 #' @examples
-#' \dontrun{
-#' if(interactive()){
-#'  #EXAMPLE1
-#'  }
-#' }
+#' color_to_hex("blue")
 #' @seealso
 #'  \code{\link[grDevices]{col2rgb}},\code{\link[grDevices]{rgb}}
 #' @rdname color_to_hex
@@ -33,17 +34,17 @@ color_to_hex <-
         }
 
 
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
-
-#' @return OUTPUT_DESCRIPTION
-#' @details DETAILS
-#' @examples
-#' \dontrun{
-#' if(interactive()){
-#'  #EXAMPLE1
-#'  }
-#' }
+#' @title
+#' Map Between Base R Colors and Hex Codes
+#'
+#' @description
+#' All colors returned by base::colors() are
+#' mapped to their corresponding hex codes.
+#'
+#' @return
+#' Vector of hexadecimal codes named by the source
+#' color name.
+#'
 #' @rdname colors_hex_map
 #' @export
 colors_hex_map <-
@@ -51,17 +52,10 @@ colors_hex_map <-
                 sapply(colors(), color_to_hex)
         }
 
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
-#' @param hex PARAM_DESCRIPTION
-#' @return OUTPUT_DESCRIPTION
-#' @details DETAILS
-#' @examples
-#' \dontrun{
-#' if(interactive()){
-#'  #EXAMPLE1
-#'  }
-#' }
+#' @title
+#' Lookup the Color Name of a Hex Code
+#'
+#'
 #' @rdname hex_to_color
 #' @export
 hex_to_color <-
@@ -71,17 +65,16 @@ hex_to_color <-
 
         }
 
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
-#' @param x PARAM_DESCRIPTION
-#' @return OUTPUT_DESCRIPTION
-#' @details DETAILS
-#' @examples
-#' \dontrun{
-#' if(interactive()){
-#'  #EXAMPLE1
-#'  }
-#' }
+#' @title
+#' Map Between Hex Codes and Base R Colors
+#'
+#' @description
+#' The inverse map of all colors returned by base::colors() are
+#' mapped to their corresponding hex codes.
+#'
+#' @return
+#' Vector of color names named by their corresponding hexadecimal code.
+#'
 #' @rdname hex_colors_map
 #' @export
 hex_colors_map <-
@@ -96,20 +89,8 @@ hex_colors_map <-
         }
 
 
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
-
-#' @return OUTPUT_DESCRIPTION
-#' @details DETAILS
-#' @examples
-#' \dontrun{
-#' if(interactive()){
-#'  #EXAMPLE1
-#'  }
-#' }
-#' @seealso
-#'  \code{\link[dplyr]{filter}}
-#'  \code{\link[rubix]{filter_at_grepl}}
+#' @title
+#' Base Colors
 #' @rdname base_colors
 #' @export
 #' @importFrom dplyr filter
@@ -144,6 +125,7 @@ base_colors <-
                         rename(base_color_x = all_colors)
 
 
+                output <-
                 left_join(
                         x = l_df,
                         y = r_df,
@@ -151,8 +133,21 @@ base_colors <-
                         keep = TRUE) %>%
                         transmute(
                                 base_color = coalesce(base_color_x, base_color_y),
-                                has_shades = !is.na(base_color_y),
-                                shade_levels = levels)
+                                color_type = ifelse(is.na(levels), "Unshaded", "Shaded"),
+                                shade_levels = ifelse(is.na(levels), 0, levels)) %>%
+                        split(.$color_type) %>%
+                        purrr::map(select, -color_type)
+
+                output$Shaded <-
+                        output$Shaded %>%
+                        split(.$shade_levels) %>%
+                        purrr::map(dplyr::select, -shade_levels)
+
+                output$Unshaded <-
+                        output$Unshaded %>%
+                        dplyr::select(-shade_levels)
+
+                return(output)
 
         }
 
@@ -162,21 +157,13 @@ base_colors <-
 # Adapted from scales::show_col. Labels are color names instead
 # of hex code.
 
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
-#' @param ... PARAM_DESCRIPTION
-#' @param labels PARAM_DESCRIPTION, Default: TRUE
-#' @param borders PARAM_DESCRIPTION, Default: NULL
-#' @param cex_label PARAM_DESCRIPTION, Default: 0.5
-#' @param ncol PARAM_DESCRIPTION, Default: NULL
-#' @return OUTPUT_DESCRIPTION
-#' @details DETAILS
-#' @examples
-#' \dontrun{
-#' if(interactive()){
-#'  #EXAMPLE1
-#'  }
-#' }
+#' @title
+#' View Colors in the Plots Pane
+#' @param ... Color names or hex codes.
+#' @param labels If TRUE, the plot is labelled with the color name. Default: TRUE
+#' @param borders Color for rectangle border(s) that is passed to `rect()`, Default: NULL
+#' @param cex_label Numeric character expansion factor; multiplied by par("cex") yields the final character size. NULL and NA are equivalent to 1.0, Default: 0.5
+#' @param ncol How wide the plot should be, Default: NULL
 #' @seealso
 #'  \code{\link[rlang]{list2}}
 #' @rdname view_colors
@@ -209,21 +196,13 @@ view_colors <-
                 }
         }
 
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
-#' @param ... PARAM_DESCRIPTION
-#' @param labels PARAM_DESCRIPTION, Default: TRUE
-#' @param borders PARAM_DESCRIPTION, Default: NULL
-#' @param cex_label PARAM_DESCRIPTION, Default: 0.5
-#' @param ncol PARAM_DESCRIPTION, Default: NULL
-#' @return OUTPUT_DESCRIPTION
-#' @details DETAILS
-#' @examples
-#' \dontrun{
-#' if(interactive()){
-#'  #EXAMPLE1
-#'  }
-#' }
+#' @title
+#' View Colors in the Plots Pane
+#' @param ... Color names or hex codes.
+#' @param labels Custom labels other than the color name.
+#' @param borders Color for rectangle border(s) that is passed to `rect()`, Default: NULL
+#' @param cex_label Numeric character expansion factor; multiplied by par("cex") yields the final character size. NULL and NA are equivalent to 1.0, Default: 0.5
+#' @param ncol How wide the plot should be, Default: NULL
 #' @seealso
 #'  \code{\link[rlang]{list2}}
 #' @rdname view_labelled_colors
@@ -234,6 +213,7 @@ view_labelled_colors <-
         function (..., labels = NULL, borders = NULL, cex_label = 0.5,
                   ncol = NULL) {
 
+                arg_length <- length(unlist(rlang::list2(...)))
                 colours <- unlist(rlang::list2(...))
 
                 n <- length(colours)
@@ -250,7 +230,7 @@ view_labelled_colors <-
                      col = colours, border = borders)
                 if (!is.null(labels)) {
 
-                        if (length(labels) == length(colours)) {
+                        if (length(labels) == arg_length) {
                                 text(x = col(colours) - 0.5,
                                      y = -row(colours) + 0.5,
                                      labels = labels,
@@ -263,17 +243,10 @@ view_labelled_colors <-
                 }
         }
 
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
-#' @param ... PARAM_DESCRIPTION
-#' @return OUTPUT_DESCRIPTION
-#' @details DETAILS
-#' @examples
-#' \dontrun{
-#' if(interactive()){
-#'  #EXAMPLE1
-#'  }
-#' }
+#' @title
+#' View Each Color and Its Shades
+#' @inheritParams view_colors
+#' @param ... Base names of colors
 #' @seealso
 #'  \code{\link[rlang]{list2}}
 #'  \code{\link[stringr]{str_remove}}
@@ -284,7 +257,9 @@ view_labelled_colors <-
 #' @importFrom stringr str_remove_all
 #' @importFrom dplyr filter select
 view_color_shades <-
-        function(...) {
+        function(...,
+                 labels = TRUE, borders = NULL, cex_label = 0.5,
+                 ncol = NULL) {
 
 
                 colours <- unlist(rlang::list2(...))
@@ -293,15 +268,13 @@ view_color_shades <-
                                        pattern = "[0-9]{1,}$")
 
                 all_colours_with_shades <-
-                        base_colors() %>%
-                        dplyr::filter(has_shades == TRUE) %>%
+                        base_colors()$Shaded$`4` %>%
                         dplyr::select(base_color) %>%
                         unlist() %>%
                         unname()
 
                 all_colours_without_shades <-
-                        base_colors() %>%
-                        dplyr::filter(has_shades == FALSE) %>%
+                        base_colors()$Unshaded %>%
                         dplyr::select(base_color) %>%
                         unlist() %>%
                         unname()
@@ -332,65 +305,6 @@ view_color_shades <-
 
                 final_colours <-
                         unname(unlist(final_colours))
-
-#
-#                 colours_with_shades2 <-
-#                         vector(mode = "list",
-#                                length = length(colours_with_shades))
-#                 names(colours_with_shades2) <-
-#                         colours_with_shades
-#
-#                 for (colour_with_shades in colours_with_shades) {
-#                         colours_with_shades2[[colour_with_shades]] <-
-#                                 c(colour_with_shades,
-#                                 sprintf("%s%s", colour_with_shades, 1:4))
-#
-#                 }
-#
-#                 colours_with_shades3 <-
-#                         unlist(colours_with_shades2) %>%
-#                         unname()
-#
-#                 if (length(colours_without_shades)>0) {
-#                         colours_without_shades2 <-
-#                                 vector(mode = "list",
-#                                        length = length(colours_without_shades))
-#                         names(colours_with_shades2) <- colours_without_shades
-#
-#
-#                         for (colour_without_shade in colours_without_shades) {
-#
-#                                 colours_without_shades2[[colour_without_shade]] <-
-#                                         c(colour_without_shade,
-#                                           rep("white", 4))
-#
-#
-#
-#                         }
-#
-#                         colours_without_shades3 <-
-#                                 unlist(colours_without_shades2) %>%
-#                                 unname()
-#
-#
-#                         final_colours <-
-#                                 c(colours_with_shades3,
-#                                   colours_without_shades3)
-#
-#
-#                 } else {
-#                         final_colours <-
-#                                 colours_with_shades3
-#
-#                 }
-#
-#                 # view_colors(final_colours,
-#                 #             ncol = 5)
-
-
-                borders <- NULL
-                cex_label <- 0.5
-                labels <- TRUE
                 n <- length(final_colours)
                 ncol <- 5
                 nrow <- ceiling(n/ncol)
@@ -421,6 +335,47 @@ view_color_shades <-
 
 
         }
+
+
+
+
+#' @title
+#' Use Colorspace Palettes for Map Assignment
+#' @description
+#' Instead of having to manually determine a map assignment,
+#' a colorspace palette can be used to auto-assign to a given
+#' vector.
+#' @param x Vector to map. All vectors are converted to character
+#' before mapping.
+#' @examples
+#' \dontrun{
+#' if(interactive()){
+#'   auto_map_to_color(x = as.character(x)) %>%
+#'   view_colors()
+#'  }
+#' }
+#' @seealso
+#'  \code{\link[colorspace]{hcl_palettes}}
+#' @rdname auto_map_to_palette
+#' @export
+#' @import colorspace
+auto_map_to_palette <-
+        function(x,
+                 colorspace_palette = "diverge_hcl",
+                 ...) {
+
+                color_assignment <-
+                        unique(as.character(x))
+
+                names(color_assignment) <-
+                        eval(rlang::parse_expr(glue::glue("colorspace::{colorspace_palette}(n = length(color_assignment))")))
+
+                fct_recode(factor(x),
+                           !!!color_assignment) %>%
+                        as.character()
+
+        }
+
 
 #' @title FUNCTION_TITLE
 #' @description FUNCTION_DESCRIPTION
@@ -455,32 +410,6 @@ assign_colors_by <-
 
 #' @title FUNCTION_TITLE
 #' @description FUNCTION_DESCRIPTION
-#' @param text PARAM_DESCRIPTION
-#' @param color PARAM_DESCRIPTION
-#' @return OUTPUT_DESCRIPTION
-#' @details DETAILS
-#' @examples
-#' \dontrun{
-#' if(interactive()){
-#'  #EXAMPLE1
-#'  }
-#' }
-#' @seealso
-#'  \code{\link[glue]{glue}}
-#' @rdname add_html_color
-#' @export
-#' @importFrom glue glue
-add_html_color <-
-        function(text,
-                 color) {
-
-                glue::glue('<span style="color: {color};">{text}</span>')
-
-
-        }
-
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
 #' @param x PARAM_DESCRIPTION
 #' @param color_assignment PARAM_DESCRIPTION
 #' @return OUTPUT_DESCRIPTION
@@ -505,32 +434,27 @@ map_colors <-
         }
 
 
-#' @title FUNCTION_TITLE
-#' @description FUNCTION_DESCRIPTION
-#' @param x PARAM_DESCRIPTION
-#' @param taken_colors PARAM_DESCRIPTION
-#' @param shade_level PARAM_DESCRIPTION, Default: 2
-#' @param seed PARAM_DESCRIPTION, Default: 100
-#' @param exclude_color_regex PARAM_DESCRIPTION
-#' @return OUTPUT_DESCRIPTION
-#' @details DETAILS
+#' @title
+#' Randomly Select Colors
+#' @param x Vector to map.
+#' @param taken_colors (Optional) Colors to exclude from selection because they have
+#' already been taken.
+#' @param shade_level Shade level from 1 to 4, Default: 2
+#' @param seed Default: 100
+#' @param exclude_color_regex (Optional) Regex to exclude a subset of colors.
 #' @examples
-#' \dontrun{
-#' if(interactive()){
-#'  #EXAMPLE1
-#'  }
-#' }
+#' auto_map_random_colors(x = c("A", "B", "C))
 #' @seealso
 #'  \code{\link[dplyr]{filter}}
-#' @rdname assign_colors
+#' @rdname auto_map_random_colors
 #' @export
 #' @importFrom dplyr filter
-assign_colors <-
+auto_map_random_colors <-
         function(x,
                  taken_colors,
                  shade_level = 2,
                  seed = 100,
-                 exclude_color_regex) {
+                 exclude_color_regex = "black|white|ivory|snow|beige|gray|grey") {
 
                 if (!missing(taken_colors)) {
                         taken_colors <-
@@ -545,13 +469,12 @@ assign_colors <-
                 }
 
                 available_base_colors <-
-                        base_colors() %>%
-                        dplyr::filter(has_shades == TRUE) %>%
+                        base_colors()$Shaded$`4` %>%
                         dplyr::filter(!(base_color %in% base_taken_colors)) %>%
                         transmute(base_color = sprintf("%s%s", base_color, shade_level)) %>%
                         unlist() %>%
                         unname() %>%
-                        grep(pattern = "black|white|ivory|snow|beige|gray|grey",
+                        grep(pattern = exclude_color_regex,
                              value = TRUE,
                              invert = TRUE)
 
@@ -588,6 +511,157 @@ assign_colors <-
                 }
 
                 output
+
+
+        }
+
+
+# From wikipedia on X11
+color_groups <-
+        list(
+        `pink` = c('mediumvioletred', 'deeppink', 'palevioletred', 'hotpink', 'lightpink', 'pink'),
+        `red` = c('darkred', 'red', 'firebrick', 'indianred', 'lightcoral', 'salmon', 'darksalmon', 'lightsalmon'),
+        `orange` = c('orangered', 'tomato', 'darkorange', 'coral', 'orange'),
+        `yellow` = c('darkkhaki', 'gold', 'khaki', 'peachpuff', 'yellow', 'palegoldenrod', 'moccasin', 'papayawhip', 'lightgoldenrodyellow', 'lemonchiffon', 'lightyellow'),
+        `brown` = c('maroon', 'brown', 'saddlebrown', 'sienna', 'chocolate', 'darkgoldenrod', 'peru', 'rosybrown', 'goldenrod', 'sandybrown', 'tan', 'burlywood', 'wheat', 'navajowhite', 'bisque', 'blanchedalmond', 'cornsilk'),
+        `green` = c('darkgreen', 'green', 'darkolivegreen', 'forestgreen', 'seagreen', 'olivedrab', 'mediumseagreen', 'limegreen', 'springgreen', 'mediumspringgreen', 'darkseagreen', 'mediumaquamarine', 'yellowgreen', 'lawngreen', 'chartreuse', 'lightgreen', 'greenyellow', 'palegreen'),
+        `cyan` = c('darkcyan', 'lightseagreen', 'cadetblue', 'darkturquoise', 'mediumturquoise', 'turquoise', 'cyan', 'aquamarine', 'paleturquoise', 'lightcyan'),
+        `blue` = c('navy', 'darkblue', 'mediumblue', 'blue', 'midnightblue', 'royalblue', 'steelblue', 'dodgerblue', 'deepskyblue', 'cornflowerblue', 'skyblue', 'lightskyblue', 'lightsteelblue', 'lightblue', 'powderblue'),
+        `purple` = c('purple', 'darkmagenta', 'darkviolet', 'darkslateblue', 'blueviolet', 'darkorchid', 'magenta', 'slateblue', 'mediumslateblue', 'mediumorchid', 'mediumpurple', 'orchid', 'violet', 'plum', 'thistle', 'lavender'),
+        `white` = c('mistyrose', 'antiquewhite', 'linen', 'beige', 'whitesmoke', 'lavenderblush', 'oldlace', 'aliceblue', 'seashell', 'ghostwhite', 'honeydew', 'floralwhite', 'azure', 'mintcream', 'snow', 'ivory', 'white'),
+        `gray` = c('black', 'darkslategray', 'dimgray', 'slategray', 'gray', 'lightslategray', 'darkgray', 'lightgray', 'gainsboro'))
+
+
+#' @title
+#' Randomly Select Colors By Color Group
+#' @param x Vector to map.
+#' @param color_group_assignment Vector named by value in X and
+#' values as one of c('pink', 'red', 'orange', 'yellow', 'brown', 'green',
+#' 'cyan', 'blue', 'purple', 'white', 'gray').
+#' @seealso
+#'  \code{\link[dplyr]{filter}}
+#' @rdname assign_colors
+#' @export
+#' @importFrom dplyr filter
+auto_map_by_color_group <-
+        function(x,
+                 color_group_assignment,
+                 seed = 100) {
+
+
+                color_assignment <- list()
+                for (i in seq_along(color_group_assignment)) {
+
+                        possible_colors <-
+                        color_groups[[color_group_assignment[i]]]
+
+                        color_assignment[[i]] <-
+                                sample(
+                                        x = possible_colors,
+                                        size = 1
+                                )
+
+
+
+                }
+                names(color_assignment) <-
+                        names(color_group_assignment)
+
+                return(color_assignment)
+
+
+        }
+
+
+#' @title
+#' Base Colors that Do Not Have Shades
+#' @rdname unshaded_base_colors
+#' @export
+#' @importFrom dplyr select
+unshaded_base_colors <-
+        function() {
+                base_colors()$Unshaded %>%
+                        dplyr::select(base_color) %>%
+                        unlist() %>%
+                        unname()
+        }
+
+
+#' @title
+#' Shaded Base Colors
+#' @rdname shaded_base_colors
+#' @export
+#' @importFrom dplyr select
+shaded_base_colors <-
+        function() {
+                base_colors()$Shaded$`4` %>%
+                        dplyr::select(base_color) %>%
+                        unlist() %>%
+                        unname()
+        }
+
+
+
+#' @title
+#' View Colors Via Regex
+#' @inheritParams base::grep
+#' @inheritParams view_colors
+#' @rdname view_colors_regex
+#' @export
+view_colors_regex <-
+        function(pattern,
+                 invert = FALSE,
+                 labels = TRUE,
+                 borders = NULL,
+                 cex_label = 0.5,
+                 ncol = NULL) {
+
+                colours <-
+                        grep(pattern = pattern,
+                             x = dplyr::bind_rows(dplyr::bind_rows(base_colors()$Shaded),
+                                                  base_colors()$Unshaded) %>%
+                                     select(base_color) %>%
+                                     unlist() %>%
+                                     unname(),
+                             value = TRUE,
+                             invert = invert)
+
+
+                view_colors(
+                        colours,
+                        labels = labels,
+                        borders = borders,
+                        cex_label = cex_label,
+                        ncol = ncol
+                )
+
+
+
+        }
+
+
+#' @title
+#' View Colors Via Regex With Shades
+#' @inheritParams base::grep
+#' @return OUTPUT_DESCRIPTION
+#' @rdname view_color_shades_regex
+#' @export
+view_color_shades_regex <-
+        function(pattern,
+                 invert = FALSE) {
+
+
+                colours <-
+                        grep(pattern = pattern,
+                             x = dplyr::bind_rows(dplyr::bind_rows(base_colors()$Shaded),
+                                                  base_colors()$Unshaded) %>%
+                                     select(base_color) %>%
+                                     unlist() %>%
+                                     unname(),
+                             value = TRUE,
+                             invert = invert)
+
+                view_color_shades(colours)
 
 
         }