+++ NEWS.md
@@ -1,6 +1,10 @@
 # easyBakeOven 1.3.0.9000    
+
 - Included NA as 'NA' for color mapping in `map_colors()`.  
 - Added `vctr_to_str_nl()`, `map_to_value()`, `view_labelled_colors()`     
+- Updated documentation for `color` functions.  
+- Added `make_tribble` and `make_list` functions.  
+
 
 
 # easyBakeOven 1.3.0